+++ src/lib/builder.ts
@@ -6,13 +6,68 @@ import { build as viteBuild } from "vite";
 import react from "@vitejs/plugin-react";
 
 import { configFileConfig, resolveConfig } from "./config.js";
-import {
-  shutdown,
-  setClientEntries,
-  getCustomModulesRSC,
-  buildRSC,
-} from "./rsc-handler.js";
+import { shutdown, setClientEntries, buildRSC } from "./rsc-handler.js";
 import { rscIndexPlugin, rscAnalyzePlugin } from "./vite-plugin-rsc.js";
+
+const createVercelOutput = (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  clientFiles: string[],
+  rscFiles: string[],
+  htmlFiles: string[]
+) => {
+  const srcDir = path.join(
+    config.root,
+    config.build.outDir,
+    config.framework.outPublic
+  );
+  const dstDir = path.join(
+    config.root,
+    config.build.outDir,
+    ".vercel",
+    "output"
+  );
+  for (const file of [...clientFiles, ...rscFiles, ...htmlFiles]) {
+    const dstFile = path.join(dstDir, "static", path.relative(srcDir, file));
+    if (!fs.existsSync(dstFile)) {
+      fs.mkdirSync(path.dirname(dstFile), { recursive: true });
+      fs.symlinkSync(path.relative(path.dirname(dstFile), file), dstFile);
+    }
+  }
+  const overrides = Object.fromEntries([
+    ...rscFiles
+      .filter((file) => !path.extname(file))
+      .map((file) => [
+        path.relative(srcDir, file),
+        { contentType: "text/plain" },
+      ]),
+    ...htmlFiles
+      .filter((file) => !path.extname(file))
+      .map((file) => [
+        path.relative(srcDir, file),
+        { contentType: "text/html" },
+      ]),
+  ]);
+  const configJson = {
+    version: 3,
+    overrides,
+  };
+  fs.mkdirSync(dstDir, { recursive: true });
+  fs.writeFileSync(
+    path.join(dstDir, "config.json"),
+    JSON.stringify(configJson, null, 2)
+  );
+};
+
+const resolveFileName = (fname: string) => {
+  for (const ext of [".js", ".ts", ".tsx", ".jsx"]) {
+    const resolvedName =
+      fname.slice(0, fname.length - path.extname(fname).length) + ext;
+    if (fs.existsSync(resolvedName)) {
+      return resolvedName;
+    }
+  }
+  return "";
+};
 
 export async function build() {
   const config = await resolveConfig("build");
@@ -24,17 +79,10 @@ export async function build() {
   );
   let entriesFile = path.join(config.root, config.framework.entriesJs);
   if (entriesFile.endsWith(".js")) {
-    for (const ext of [".js", ".ts", ".tsx", ".jsx"]) {
-      const tmp = entriesFile.slice(0, -3) + ext;
-      if (fs.existsSync(tmp)) {
-        entriesFile = tmp;
-        break;
-      }
-    }
+    entriesFile = resolveFileName(entriesFile) || entriesFile;
   }
   const require = createRequire(import.meta.url);
 
-  const customModules = await getCustomModulesRSC();
   const clientEntryFileSet = new Set<string>();
   const serverEntryFileSet = new Set<string>();
   await viteBuild({
@@ -58,7 +106,6 @@ export async function build() {
       rollupOptions: {
         input: {
           entries: entriesFile,
-          ...customModules,
         },
       },
     },
@@ -88,7 +135,6 @@ export async function build() {
           entries: entriesFile,
           ...clientEntryFiles,
           ...serverEntryFiles,
-          ...customModules,
         },
         output: {
           banner: (chunk) => {
@@ -103,10 +149,13 @@ export async function build() {
             return code;
           },
           entryFileNames: (chunkInfo) => {
-            if (chunkInfo.name === "entries" || customModules[chunkInfo.name]) {
-              return "[name].js";
+            if (
+              clientEntryFiles[chunkInfo.name] ||
+              serverEntryFiles[chunkInfo.name]
+            ) {
+              return "assets/[name].js";
             }
-            return "assets/[name].js";
+            return "[name].js";
           },
         },
       },
@@ -166,7 +215,7 @@ export async function build() {
   );
   await setClientEntries(absoluteClientEntries);
 
-  await buildRSC();
+  const buildOutput = await buildRSC();
 
   const origPackageJson = require(path.join(config.root, "package.json"));
   const packageJson = {
@@ -182,6 +231,22 @@ export async function build() {
   fs.writeFileSync(
     path.join(config.root, config.build.outDir, "package.json"),
     JSON.stringify(packageJson, null, 2)
+  );
+
+  // https://vercel.com/docs/build-output-api/v3
+  // So far, only static sites are supported.
+  createVercelOutput(
+    config,
+    clientBuildOutput.output.map(({ fileName }) =>
+      path.join(
+        config.root,
+        config.build.outDir,
+        config.framework.outPublic,
+        fileName
+      )
+    ),
+    buildOutput.rscFiles,
+    buildOutput.htmlFiles
   );
 
   await shutdown();