+++ src/lib/rsc-handler.ts
@@ -13,8 +13,9 @@ export type RenderInput<Props extends {} = {}> = {
   args?: unknown[] | undefined;
 };
 
-type CustomModules = {
-  [name: string]: string;
+export type BuildOutput = {
+  rscFiles: string[];
+  htmlFiles: string[];
 };
 
 export type MessageReq =
@@ -24,26 +25,15 @@ export type MessageReq =
       type: "setClientEntries";
       value: "load" | Record<string, string>;
     }
-  | {
-      id: number;
-      type: "render";
-      input: RenderInput;
-    }
-  | {
-      id: number;
-      type: "getCustomModules";
-    }
-  | {
-      id: number;
-      type: "build";
-    };
+  | { id: number; type: "render"; input: RenderInput }
+  | { id: number; type: "build" };
 
 export type MessageRes =
   | { type: "full-reload" }
   | { id: number; type: "buf"; buf: ArrayBuffer; offset: number; len: number }
   | { id: number; type: "end" }
   | { id: number; type: "err"; err: unknown }
-  | { id: number; type: "customModules"; modules: CustomModules };
+  | { id: number; type: "buildOutput"; output: BuildOutput };
 
 const messageCallbacks = new Map<number, (mesg: MessageRes) => void>();
 
@@ -113,29 +103,12 @@ export function renderRSC(input: RenderInput): Readable {
   return passthrough;
 }
 
-export function getCustomModulesRSC(): Promise<CustomModules> {
-  return new Promise<CustomModules>((resolve, reject) => {
+export function buildRSC(): Promise<BuildOutput> {
+  return new Promise((resolve, reject) => {
     const id = nextId++;
     messageCallbacks.set(id, (mesg) => {
-      if (mesg.type === "customModules") {
-        resolve(mesg.modules);
-        messageCallbacks.delete(id);
-      } else if (mesg.type === "err") {
-        reject(mesg.err);
-        messageCallbacks.delete(id);
-      }
-    });
-    const mesg: MessageReq = { id, type: "getCustomModules" };
-    worker.postMessage(mesg);
-  });
-}
-
-export function buildRSC(): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    const id = nextId++;
-    messageCallbacks.set(id, (mesg) => {
-      if (mesg.type === "end") {
-        resolve();
+      if (mesg.type === "buildOutput") {
+        resolve(mesg.output);
         messageCallbacks.delete(id);
       } else if (mesg.type === "err") {
         reject(mesg.err);