+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+## [0.11.2] - 2023-06-03
+### Changed
+- fix(examples): config extension in script #56
+- vercel build output api #53
+
+## [0.11.1] - 2023-06-01
+### Changed
+- fix: rscPrefix config without trailing slash #50
+- feat: generalized router #55
+
 ## [0.11.0] - 2023-05-25
 ### Changed
 - fix: make getBuilder optional #41